pyproject.toml
[build-system]
requires = ["setuptools>=64"]
build-backend = "setuptools.build_meta"

[project]
name = "xlsx2parquet-tools"
version = "0.0.1"
dependencies = [
    "pandas",
    "pyarrow",
    "openpyxl",
    "streamlit",
    "groq",
    "python-dotenv"
]

[tool.setuptools.packages.find]
where = ["src"]